requirements.txt
flask
mysql-connector-python
requests
meteostat==2.1.3
pandas
numpy==1.24.4
flask-cors